Port: 0.4.1 (link100 wave 20-21)

+++ src/Vec3_LslInPlace.c
@@ -1,7 +1,7 @@
 int *Vec3_LslInPlace(int *v, int sh)
 {
     v[0] <<= sh;
-    *(int *)(((long long)(int)(v + 1))) <<= sh;
-    *(int *)(((long long)(int)(v + 2))) <<= sh;
+    *(int *)(v + 1) <<= sh;
+    *(int *)(v + 2) <<= sh;
     return v;
 }
