+++ src/func_020355c8.c
@@ -1,4 +1,4 @@
 void func_020355c8(char *self)
 {
-    *(unsigned int *)(((long long)(int)(self + 0x10))) &= ~0x800;
+    *(unsigned int *)(self + 0x10) &= ~0x800;
 }
