+++ src/func_02046e28.c
@@ -7,7 +7,7 @@ struct Model { u8 pad0[4]; struct Material *materials; };
 
 extern void Crash(void);
 
-#define LAUNDER(p) ((volatile u32 *)(int)(((long long)(int)(p))))
+#define LAUNDER(p) ((volatile u32 *)(int)(p))
 
 void func_02046e28(struct Model *model, struct MatAnimData *anim, short frame)
 {
