+++ src/func_02059f58.c
@@ -5,7 +5,7 @@ void func_02059f58(int channel)
 {
     unsigned int saved = _ZN3IRQ7DisableEv();
     int idx = channel * 6 + 5;
-    volatile unsigned short *reg = (volatile unsigned short *)(((long long)(int)(0x40000b0 + idx * 2)));
+    volatile unsigned short *reg = (volatile unsigned short *)(0x40000b0 + idx * 2);
     *reg &= ~0x3a00;
     *reg &= ~0x8000;
     (void)*reg;
