+++ src/func_0205a064.c
@@ -17,13 +17,13 @@ void func_0205a064(u32 ch, u32 src, u32 dst, u32 size, void (*cb)(u32), u32 cbAr
     if (cb != 0)
     {
         func_02056e98(ch, (u32)cb, cbArg);
-        reg = (u32 *)(int)(((long long)(int)(0x040000e0 + ch * 4)));
+        reg = (u32 *)(int)(0x040000e0 + ch * 4);
         *reg = dst;
         DMAStartTransfer(ch, (u32)reg, src, (size >> 2) | 0xc5000000);
     }
     else
     {
-        reg = (u32 *)(int)(((long long)(int)(0x040000e0 + ch * 4)));
+        reg = (u32 *)(int)(0x040000e0 + ch * 4);
         *reg = dst;
         DMAStartTransfer(ch, (u32)reg, src, (size >> 2) | 0x85000000);
     }
