+++ src/func_ov006_02114fec.c
@@ -1,5 +1,5 @@
 void func_ov006_02114fec(char *p)
 {
-    int *a = (int *)(((long long)(int)(p + 0x5980)));
+    int *a = (int *)(p + 0x5980);
     *a = *a + 1;
 }
