+++ src/func_ov006_02115060.c
@@ -1,5 +1,5 @@
 void func_ov006_02115060(char *p)
 {
-    int *a = (int *)(((long long)(int)(p + 0x5964)));
+    int *a = (int *)(p + 0x5964);
     *a = *a + 1;
 }
