+++ src/func_ov007_020b7708.c
@@ -19,6 +19,6 @@ void func_ov007_020b7708(int idx, void* pos)
         struct PtclSys* s;
         data_ov007_02103450[idx] = _ZN8Particle7Manager9AddSystemEiR7Vector3(data_ov007_0210344c, 0, pos);
         s = data_ov007_02103450[idx];
-        *(u32*)(((long long)(int)((char*)s + 0x1c))) |= 2;
+        *(u32*)((char*)s + 0x1c) |= 2;
     }
 }
